Rename `rustc_ast_lowering/src/errors.rs` into `rustc_ast_lowering/src/diagnostics.rs`
